+++ packages/apollo-server-core/src/plugin/landingPage/default/index.ts
@@ -14,6 +14,13 @@ export interface ApolloServerPluginLandingPageDefaultBaseOptions {
    * landing page.
    */
   footer?: boolean;
+  /**
+   * Folks can configure their landing page to link to Studio Explorer with a
+   * document, variables and headers loaded in the UI.
+   */
+  document?: string;
+  variables?: Record<string, string>;
+  headers?: Record<string, string>;
   // For Apollo use only.
   __internal_apolloStudioEnv__?: 'staging' | 'prod';
 }
@@ -37,6 +44,9 @@ interface LandingPageConfig {
   graphRef?: string | undefined;
   isProd?: boolean;
   apolloStudioEnv?: 'staging' | 'prod';
+  document?: string;
+  variables?: Record<string, string>;
+  headers?: Record<string, string>;
   footer?: boolean;
 }